toil compatible variant recalibration changes
Instead of passing an array of values for the settings
to use with resource files we hard code them in the tool.
This results in two tools instead of one.
This is to work around an error when the subworkflow
was passing an array of settings to a tool:
`TypeError: tuple indices must be integers, not unicode`
See https://github.com/DataBiosphere/toil/issues/2814